add comment, fix. Add a new property - avatar for the user.
